Add test
It is super heavy atm, as it instantiates the whole gui, but better than not having a test. In the future some of the infrastructure could be refactored out similar to pyiron_base's TestWithCleanProject.
I also had to make the package setting variable; there should be a more elegant solution for this going forward as well.