Improve abstraction
This is a hideous commit, the scope of which belies my continued lack of familiarity with ipywidgets and guis. The one redeeming factor is that the Tab class actually has a bug that wasn't obvious in the old formulation, and finding and working around this is part of why this got so ugly. Whenever the number of scripts change, things are a bit clunky as I totally rewrite the canvases; this might need fixing later, or it might be OK since that's not a common operation and it's still only fractional seconds for small graphs.