Nevermind, suddenly canvas is working
What a terrible bug. I didn't even rebuild the environment, binder still had it in cache, but now today the session for commit c505020 worked fine. NGLView is now giving an error, which it doesn't at the original mybinder link before I started working on the project, and there's no fancy postBuild or anything there, so who knows what is causing that. However, Lammps won't actualy run because no executable exists. This is pretty sensible since lammps is not an actual dependency of atomistics.