Use Mambaforge
Per [the docs](https://docs.readthedocs.io/en/stable/guides/conda.html) and also in line with the other pyiron modules. In the readthedocs build logs I can explicitly see that nbsphinx is not getting added in the env update, I suspect but am not certain this had to do with specifying python rather than conda in the build.