Reorder matrix
condamerge.py is failing with ImportError: No module named yaml, even though it is working fine in both the docs and notebooks workflows. The only difference I can see is that unittests start on mac while the other two use linux. Let's reorder the matrix so linux runs first and see if it gets through these before breaking at mac.

This is very strange, because until recently condamerge was also working on mac. Ahaaaa, but until recently we were in a logic branch that just used bash's cp instead of the condamerge script for the unittests! Perhaps this script has never worked on mac.