Recursively close
It smells to do this, and what's more _it doesn't do anything_. Or at least it doesn't do everything, as deleting the main GUI object after instantiation still leaves all the widgets around in the registry. There must still be a reference to them somewhere I'm not seeing, so they are not allowed to be purged. Maybe once that reference is found and eliminated this recursive closing will be important.