Match the list signature from colab
My local python installation allows a `seq` kwarg to list, which caused an error on my binder. I'm still not 100% if this is some pycharm magic (unlikely) or a change between 3.8 and 3.11 (likely). Interestingly, even the mybinder help for `list??` erroneously states `Init signature: list(iterable=(), /)`. In fact, no kwargs are possible. Only a single (iterable) arg is allowable, but I'm just going to follow the colab `list??` and allow the full generic signature and let users hit the error themselves.