Lock dependencies (#216)
* Pin the key pyiron dependencies to latest

And let's see what happens

* Add aimsgb

* Remove maggma

In the newer pyiron_atomistics dependency this should now be resolved

* Freeze dependencies

Locally this created an environment amenable to all demo notebooks. Let's check it on MyBinder.

* Pin the key pyiron dependencies to latest

And let's see what happens

* Add aimsgb

* Remove maggma

In the newer pyiron_atomistics dependency this should now be resolved

* Freeze dependencies

Locally this created an environment amenable to all demo notebooks. Let's check it on MyBinder.

* [dependabot skip] Update env file

* Use python 11 instead of 12

* Exclude python 3.8

* [dependabot skip] Update env file

* Add missing dependencies

This was simply a transcription error

* [dependabot skip] Update env file

* Synchronzie setup.py

* Revise environment

* [dependabot skip] Update env file

* Fix typo in the jupyterlab dependence

* [dependabot skip] Update env file

* Fix the same typo for matplotlib

* [dependabot skip] Update env file

* Relax the pip dependency for pyiron_gui

* Bump the timeout for windows to see if that fixes things

---------

Co-authored-by: pyiron-runner <pyiron@mpie.de>